% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate weights for explanation model}
\usage{
calculate_weights(simulated_dataset, explained_instance, kernel)
}
\arguments{
\item{simulated_dataset}{Dataset simulated by sample_locally function.}

\item{explained_instance}{Instance to be explained.}

\item{kernel}{Chosen kernel function.}
}
\value{
Numeric vector of weights for each row in simulated dataset.
}
\description{
Calculate weights for explanation model
}
