% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{check_conditions}
\alias{check_conditions}
\title{Check if data, explained instance and size make sense.}
\usage{
check_conditions(data, explained_instance, size)
}
\arguments{
\item{data}{Data frame from which observations will be sampled.}

\item{explained_instance}{Instance around which points will be sampled.}

\item{size}{Number of observation in simulated dataset}
}
\value{
Produces an error if any of conditions aren't met.
}
\description{
Check if data, explained instance and size make sense.
}
