% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_for_na}
\alias{check_for_na}
\title{Alert user if NAs are present}
\usage{
check_for_na(data, explained_instance)
}
\arguments{
\item{data}{Data frame from which observations are generated by sample_locally.}

\item{explained_instance}{A row in an original data frame (as a data.frame).}
}
\value{
Warning message
}
\description{
Alert user if NAs are present
}
