% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimateWhiteBoxRegLin}
\alias{estimateWhiteBoxRegLin}
\title{Estimate linear regression white box model}
\usage{
estimateWhiteBoxRegLin(originalObservation, simulatedData, explainedVar,
  kernel = "gaussian", sigma = 1)
}
\arguments{
\item{originalObservation}{observation (vector) around which local
exploration is being done}

\item{simulatedData}{d.f returned by generateNeighbourhood function}

\item{explainedVar}{vector of predictions from black box model}

\item{kernel}{kernel function used to calculate weights}

\item{sigma}{parameter used in weights calculation}
}
\value{
lm object
}
\description{
Estimate linear regression white box model
}
