% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{euclidean_kernel}
\alias{euclidean_kernel}
\title{LIME kernel equal to the inverse of euclidean distance.}
\usage{
euclidean_kernel(explained_instance, simulated_instance)
}
\arguments{
\item{explained_instance}{explained instance}

\item{simulated_instance}{new observation}
}
\value{
numeric
}
\description{
LIME kernel equal to the inverse of euclidean distance.
}
