% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{LIME kernel from the original article with sigma = 1.}
\usage{
gaussian_kernel(explained_instance, simulated_instance)
}
\arguments{
\item{explained_instance}{explained instance}

\item{simulated_instance}{new observation}
}
\value{
numeric
}
\description{
LIME kernel from the original article with sigma = 1.
}
