% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_code.R
\name{generate_neighbourhood}
\alias{generate_neighbourhood}
\title{LIME: sampling for local exploration}
\usage{
generate_neighbourhood(data, explained_instance, size)
}
\arguments{
\item{data}{Data frame from which observations will be generated.}

\item{explained_instance}{A row in an original data frame (as a data.frame).}

\item{size}{Number of observations to be generated.}
}
\value{
data.frame
}
\description{
LIME: sampling for local exploration
}
