% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_methods.R
\name{generate_neighbourhood2}
\alias{generate_neighbourhood2}
\title{LIME: sampling for local exploration by changing one value per observation.}
\usage{
generate_neighbourhood2(data, explained_instance, size, fixed_variables,
  seed = NULL)
}
\arguments{
\item{data}{Data frame from which observations will be generated.}

\item{explained_instance}{A row in an original data frame (as a data.frame).}

\item{size}{Number of observations to be generated.}

\item{fixed_variables}{Names of column which will not be changed while sampling.}

\item{seed}{Seed to set before sampling. If NULL, results will not be reproducible.}
}
\value{
data.frame
}
\description{
LIME: sampling for local exploration by changing one value per observation.
}
