% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNeighbourhood.R
\name{generateNeighbourhood}
\alias{generateNeighbourhood}
\title{LIME: sampling for local exploration}
\usage{
generateNeighbourhood(observation, noOfNeighbours, originalDataFrame)
}
\arguments{
\item{observation}{a number of row in an original data frame}

\item{noOfNeighbours}{number of observations to be generated}

\item{originalDataFrame}{d.f from which observations will be generated
excluding the explained variable}
}
\description{
LIME: sampling for local exploration
}
