% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{give_predictions2}
\alias{give_predictions2}
\title{Add predictions to generated dataset.}
\usage{
give_predictions2(data, black_box, explained_var, similar,
  predict_function, hyperpars = list(), ...)
}
\arguments{
\item{data}{Original data frame used to generate new dataset.}

\item{black_box}{String with mlr signature of a learner or a model with predict interface.}

\item{explained_var}{Name of a column with the variable to be predicted.}

\item{similar}{Dataset created for local exploration.}

\item{predict_function}{Either a "predict" function that returns a vector of the
same type as response or custom function that takes a model as a first argument,
new data used to calculate predictions as a second argument called "newdata"
and returns a vector of the same type as response.
Will be used only if a model object was provided in the black_box argument.}

\item{hyperpars}{Optional list of (hyper)parameters to be passed to mlr::makeLearner.}

\item{...}{Additional parameters to be passed to predict function.}
}
\value{
A list that consists of black box model object and predictions.
}
\description{
Add predictions to generated dataset.
}
