% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dalex_shortcut.R, R/live.R
\docType{package}
\name{live}
\alias{live}
\alias{live-package}
\title{Fit local model around the observation: shortcut for DALEX explainer objects}
\usage{
live(explainer, observation, target_variable_name, n_new_obs,
  local_model = "regr.lm", select_variables = F,
  predict_type = "response", kernel_type = gaussian_kernel, ...)
}
\arguments{
\item{explainer}{a model to be explained, preprocessed by the DALEX::explain function}

\item{observation}{a new observation for which predictions need to be explained}

\item{target_variable_name}{name of the response variablea as a character}

\item{n_new_obs}{Number of observation in the simulated dataset}

\item{local_model}{Character specyfing mlr learner to be used as a local model}

\item{select_variables}{If TRUE, variable selection will be performed while
fitting the local linear model}

\item{predict_type}{Argument passed to mlr::makeLearner() argument "predict.type"
while fitting the local model. Defaults to "response"}

\item{kernel_type}{Function which will be used to calculate distances from
simulated observation to explained instance}

\item{...}{Arguments to be passed to sample_locally2 function}
}
\description{
This package aims to help locally fit and visualize interpretable models similarly to LIME methodology.
Interface provided by mlr package is used. Tools are provided to create a simulated dataset of
similar observations, fit a chosen white box models (GLM and CART in particular) and visualize
them. The methodology is based on Tulio Ribeiro, Singh, Guestrin (2016) <doi:10.1145/2939672.2939778>.
More details can be found in Staniak, Biecek (2018) https://arxiv.org/abs/1804.01955.
}
\section{Important functions}{

\code{\link{sample_locally2}} generates a dataset that will be used for local exploration.
\code{\link{add_predictions2}} adds black box model predictions to simulated dataset.
\code{\link{fit_explanation2}} fits a chosen white box model to simulated dataset.
\code{\link{plot}} visualizes fitted model.
}

\section{Example datasets}{

\code{wine} Data on wine quality taken from
Modeling wine preferences by data mining from physicochemical properties
}

