% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{live_shiny}
\alias{live_shiny}
\title{Function that starts a Shiny app which helps use LIVE.}
\usage{
live_shiny(train_data, black_box_model, target,
  explained_data = train_data)
}
\arguments{
\item{train_data}{dataset from which observations will be sampled.}

\item{black_box_model}{Pre-trained  model with predict interface.}

\item{target}{character, name of the response variable.}

\item{explained_data}{Data frame with predictions to explain.}
}
\value{
shiny app
}
\description{
Function that starts a Shiny app which helps use LIVE.
}
