% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_methods.R
\name{normal_neighbourhood}
\alias{normal_neighbourhood}
\title{LIME: sampling for local exploration using normal distribution.}
\usage{
normal_neighbourhood(data, explained_instance, size, fixed_variables, ...)
}
\arguments{
\item{data}{Data frame from which observations will be generated.}

\item{explained_instance}{A row in an original data frame (as a data.frame).}

\item{size}{Number of observations to be generated.}

\item{fixed_variables}{Names of column which will not be changed while sampling.}

\item{...}{Mean and covariance matrix for the normal distribution.}
}
\value{
data.frame
}
\description{
LIME: sampling for local exploration using normal distribution.
}
