% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{permutation_neighbourhood}
\alias{permutation_neighbourhood}
\title{LIME: sampling for local exploration by permuting all columns.}
\usage{
permutation_neighbourhood(data, explained_instance, size, fixed_variables)
}
\arguments{
\item{data}{Data frame from which observations will be generated.}

\item{explained_instance}{A row in an original data frame (as a data.frame).}

\item{size}{Number of observations to be generated.}

\item{fixed_variables}{Names of column which will not be changed while sampling.}
}
\value{
data frame
}
\description{
LIME: sampling for local exploration by permuting all columns.
}
