% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.live}
\alias{plot.live}
\title{Plotting white box models.}
\usage{
\method{plot}{live}(liveObject)
}
\arguments{
\item{liveObject}{Object returned by whiteBoxify function.}
}
\value{
plot
}
\description{
Plotting white box models.
}
