% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.live_explainer}
\alias{plot.live_explainer}
\title{Plotting white box models.}
\usage{
\method{plot}{live_explainer}(x, ..., type = NULL)
}
\arguments{
\item{x}{List returned by fit_explanation function.}

\item{...}{Additional parameters.}

\item{type}{Chr, "forest" or "waterfall" depending
on which type of plot is to be created.
if lm/glm model is used as interpretable approximation.}
}
\value{
plot (ggplot2 or base)
}
\description{
Plotting white box models.
}
\examples{
\dontrun{
# Forest plot for regression
plot(fitted_explanation1, type = "forest")
# Waterfall plot
plot(fitted_explanation1, type = "waterfall")
# Plot decision tree
plot(fitted_explanation2)
}

}
