% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_perimp.R
\name{plot.local_permutation_importance}
\alias{plot.local_permutation_importance}
\title{Plot local permutation importance}
\usage{
\method{plot}{local_permutation_importance}(x, ...)
}
\arguments{
\item{x}{Object of class local_permutation_importance}

\item{...}{Optional arguments, currently ignored}
}
\value{
ggplot2 object
}
\description{
Plot local permutation importance
}
