% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_interface.R
\name{plot_explanation2}
\alias{plot_explanation2}
\title{Plotting white box models.}
\usage{
plot_explanation2(explained_model, regr_plot_type = NULL)
}
\arguments{
\item{explained_model}{List returned by fit_explanation function.}

\item{regr_plot_type}{Chr, "forest" or "waterfall" depending
on which type of plot is to be created.
if lm/glm model is used as interpretable approximation.}
}
\value{
plot (ggplot2 or base)
}
\description{
DEPRECATED. Please refer to the generic plot function.
}
\examples{
\dontrun{
# Forest plot for regression
plot_explanation(fitted_explanation1, "forest", wine[5, ])
# Waterfall plot
plot_explanation(fitted_explanation1, "waterfall", wine[5, ])
# Plot decision tree
plot_explanation(fitted_explanation2)
}

}
