% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{plot_regression}
\alias{plot_regression}
\title{Watefall plot or forestplot for lm/glm explanations.}
\usage{
plot_regression(plot_type, fitted_model, explained_instance)
}
\arguments{
\item{plot_type}{}
}
\description{
Watefall plot or forestplot for lm/glm explanations.
}
