% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_regression2}
\alias{plot_regression2}
\title{Waterfall plot or forestplot for lm/glm explanations.}
\usage{
plot_regression2(plot_type, fitted_model, explained_instance, scale = NULL)
}
\arguments{
\item{plot_type}{Chr, "forest" or "waterfall" depending
on which type of plot is to be created.}

\item{fitted_model}{glm or lm object.}

\item{explained_instance}{Observation around which model was fitted.}

\item{scale}{Only for classification problems, "logit" or "probability".}
}
\value{
plot (ggplot2 or lattice)
}
\description{
Waterfall plot or forestplot for lm/glm explanations.
}
