% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotWhiteBox}
\alias{plotWhiteBox}
\title{Plotting white box models.}
\usage{
plotWhiteBox(whiteBox, observation = NULL)
}
\arguments{
\item{whiteBox}{object returned by mlr::train function.}

\item{observation}{Observation around which model was fitted.}
}
\value{
plot
}
\description{
Plotting white box models.
}
