% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{prepareForestplot}
\alias{prepareForestplot}
\title{Draw a forest plot with proper annotations.}
\usage{
prepareForestplot(coefficients, explained_instance)
}
\arguments{
\item{coefficients}{Matrix of regression coefficients (from summary).}

\item{explained_instance}{Instance which is being explained (as a row).}
}
\description{
Draw a forest plot with proper annotations.
}
