% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.live_explainer}
\alias{print.live_explainer}
\title{Generic print function for live explainer}
\usage{
\method{print}{live_explainer}(x, ...)
}
\arguments{
\item{x}{Object created using fit_explanation function}

\item{...}{other arguments}
}
\description{
Generic print function for live explainer
}
