% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.live_explorer}
\alias{print.live_explorer}
\title{Generic print function for class live_explorer}
\usage{
\method{print}{live_explorer}(x, ...)
}
\arguments{
\item{x}{Object created by sample_locally function or add_predictions function}

\item{...}{Other arguments}
}
\description{
Generic print function for class live_explorer
}
