% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_perimp.R
\name{print.local_permutation_importance}
\alias{print.local_permutation_importance}
\title{Print method for local_permutation_importance class}
\usage{
\method{print}{local_permutation_importance}(x, ...)
}
\arguments{
\item{x}{Object of class local_permutation_importance}

\item{...}{Optional arguments, currently ignored}
}
\description{
Print method for local_permutation_importance class
}
