% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNeighbourhood.R
\name{replace_items}
\alias{replace_items}
\title{Replace one element of a vector with a random element from given data frame.}
\usage{
replace_items(chosen_vars, data, new_data)
}
\arguments{
\item{chosen_vars}{Columns to sample from [as positions/numbers].}

\item{data}{Data frame to sample from.}

\item{new_data}{Data frame in which the replacement is done.}
}
\value{
data.frame
}
\description{
Replace one element of a vector with a random element from given data frame.
}
