% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{replace_items_by_rows}
\alias{replace_items_by_rows}
\title{Replace one element of each row with a random element from given data frame.}
\usage{
replace_items_by_rows(chosen_vars, data, target)
}
\arguments{
\item{chosen_vars}{Columns to sample from [as positions/numbers].}

\item{data}{Data frame to sample from.}

\item{target}{Data frame in which the replacement is done.}
}
\value{
data.frame
}
\description{
Replace one element of each row with a random element from given data frame.
}
