% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNeighbourhood.R
\name{replaceItems}
\alias{replaceItems}
\title{Replace one element of a vector with a random element from given d.f}
\usage{
replaceItems(chosenVariables, originalDataFrame, newDataFrame)
}
\arguments{
\item{chosenVariables}{numbers of columns from d.f to sample from}

\item{originalDataFrame}{d.f to sample from}

\item{newDataFrame}{d.f in which the replacement is done}
}
\value{
data.frame
}
\description{
Replace one element of a vector with a random element from given d.f
}
