% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{select_variables}
\alias{select_variables}
\title{Select variables for explanation model.}
\usage{
select_variables(source_data, target, response_family)
}
\arguments{
\item{source_data}{Simulated dataset.}

\item{target}{Name of the response variable.}

\item{response_family}{Name of distribution family to be used in lasso/glm fit.}
}
\value{
Character vector of names of selected variables
}
\description{
Select variables for explanation model.
}
