% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_interface.R
\name{set_constant_dates}
\alias{set_constant_dates}
\title{Set date values to one value}
\usage{
set_constant_dates(data, explained_instance)
}
\arguments{
\item{data}{Data frame to change.}

\item{explained_instance}{1-row data frame with instance of interest.}
}
\description{
Set date values to one value
}
