% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_constant_variables}
\alias{set_constant_variables}
\title{Set date values to one value}
\usage{
set_constant_variables(data, explained_instance, col_names)
}
\arguments{
\item{data}{Data frame to change.}

\item{explained_instance}{Instance that will be explained.}

\item{col_names}{Names of columns to be fixed}
}
\description{
Set date values to one value
}
