% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{shapley_kernel}
\alias{shapley_kernel}
\title{LIME kernel from the article about Shapley values}
\usage{
shapley_kernel(explained_instance, simulated_instance)
}
\arguments{
\item{explained_instance}{explained instance}

\item{simulated_instance}{new observation}
}
\value{
numeric
}
\description{
LIME kernel from the article about Shapley values
}
