% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatesimilar.R
\name{simulate_similar}
\alias{simulate_similar}
\title{Generate dataset for white box model based on black box model.}
\usage{
simulate_similar(data, explained_instance, explained_var, black_box, size,
  standardise = FALSE, predict_function = predict, ...)
}
\arguments{
\item{data}{Data frame from which new dataset will be simulated.}

\item{explained_instance}{One row data frame with the same variables 
as in data argument. Local exploration will be performed around this observation.}

\item{explained_var}{Name of a column with the variable to be predicted.}

\item{black_box}{String with mlr signature of a learner or a model with predict interface.}

\item{size}{Number of observations is a simulated dataset.}

\item{standardise}{If TRUE, numerical variables will be scaled to have mean 0, var 1.}

\item{predict_function}{Either a "predict" function that returns a vector of the
same type as response or custom function that takes a model as a first argument,
new data used to calculate predictions as a second argument called "newdata"
and returns a vector of the same type as respone. 
Will be used only if a model object was provided in the black_box argument.}

\item{...}{Additional parameters to be passed to makeRegrTask function.}
}
\value{
list
}
\description{
Generate dataset for white box model based on black box model.
}
