% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatesimilar.R
\name{simulateSimilar}
\alias{simulateSimilar}
\title{Generate dataset for white box model based on black box model.}
\usage{
simulateSimilar(data, newData, explainedVar, blackBox, noOfNeighbours,
  standardise = FALSE, predictionFunction = predict, ...)
}
\arguments{
\item{data}{D.f with variables, from which new dataset will be simulated.}

\item{newData}{One-row d.f with the same variables as in originalDataset argument,
local exploration will be performed around this observation.}

\item{explainedVar}{Name of a column with the variable to be predicted.}

\item{blackBox}{String with mlr signature of a learner or a model
with predict interface.}

\item{noOfNeighbours}{Number of similar observations to simulate.}

\item{standardise}{If TRUE, numerical variables will be scaled to have mean 0, var 1.}

\item{predictionFunction}{Either a "predict" function that returns a vector of the
same type as response or custom function that takes a model as a first argument,
new data to base predictions on as a second argument and returns a vector
of the same type as respone. Will be used only if a model object was provided
in the blackBox argument.}

\item{...}{Additional parameters to be passed to makeRegrTask function.}
}
\value{
list
}
\description{
Generate dataset for white box model based on black box model.
}
