% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatesimilar.R
\name{trainWhiteBox}
\alias{trainWhiteBox}
\title{Fit white box model to the simulated data.}
\usage{
trainWhiteBox(liveObject, whiteBox, selection = FALSE, maxDepth = 0)
}
\arguments{
\item{liveObject}{List return by simulateSimilar function.}

\item{whiteBox}{String, learner name recognized by mlr package.}

\item{selection}{If TRUE, variable selection based on AIC will be performed.}

\item{maxDepth}{maximum depth of a tree (when decision tree is used).}
}
\value{
mlr object returned by train function.
}
\description{
Fit white box model to the simulated data.
}
