% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNeighbourhood.R
\name{walk_through_vars}
\alias{walk_through_vars}
\title{Change value of one variable all rows.}
\usage{
walk_through_vars(data, new_data, steps)
}
\arguments{
\item{data}{Data frame from which observations will be generated.}

\item{new_data}{Data frame created by generate_neighbourhood function.}

\item{steps}{Number of variables to change.}
}
\value{
data.frame
}
\description{
Change value of one variable all rows.
}
