% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNeighbourhood.R
\name{walkThroughVariables}
\alias{walkThroughVariables}
\title{Change value of one variable all rows.}
\usage{
walkThroughVariables(originalDataFrame, newDataFrame, steps)
}
\arguments{
\item{originalDataFrame}{d.f from which observations will be generated
excluding the explained variable}

\item{newDataFrame}{d.f created by generateNeighbourhood function}

\item{steps}{Number of variables to change}
}
\description{
Change value of one variable all rows.
}
