% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteboxify.R
\name{whiteboxify}
\alias{whiteboxify}
\title{Create white box model based on a black box.}
\usage{
whiteboxify(data, newData, explainedVar, blackBox, whiteBox, noOfNeighbours,
  standardise = FALSE, ...)
}
\arguments{
\item{data}{D.f with variables, from which new dataset will be simulated.}

\item{newData}{One-row d.f with the same variables as in originalDataset argument,
local exploration will be performed around this observation.}

\item{explainedVar}{Name of a column with the variable to be predicted.}

\item{blackBox}{String with mlr signature of a learner or a model
with predict interface.}

\item{whiteBox}{String, "reg" for linear regression or "dtree" for decision tree.}

\item{noOfNeighbours}{Number of similar observations to simulate.}

\item{standardise}{If TRUE, numerical variables will be scaled to have mean 0, var 1.}

\item{...}{Additional parameters to be passed to makeRegrTask function.}
}
\value{

}
\description{
Create white box model based on a black box.
}
