% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/live.R
\docType{data}
\name{wine}
\alias{wine}
\title{Red wine characteristics and quality.}
\format{Data frame with 1599 rows and 12 columns.}
\usage{
wine
}
\description{
Popular dataset related to wine samples from north Portugal.
}
\references{
P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis.
Modeling wine preferences by data mining from physicochemical properties.
In Decision Support Systems, Elsevier, 47(4):547-553, 2009.
}
\keyword{datasets}
