% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{add_predictions}
\alias{add_predictions}
\title{Add black box predictions to generated dataset}
\usage{
add_predictions(to_explain, black_box_model, predict_fun = predict, ...)
}
\arguments{
\item{to_explain}{List return by sample_locally function.}

\item{black_box_model}{String with mlr signature of a learner or a model with predict interface.}

\item{...}{Additional parameters to be passed to predict function.}

\item{predict_function}{Either a "predict" function that returns a vector of the
same type as response or custom function that takes a model as a first argument,
new data used to calculate predictions as a second argument called "newdata"
and returns a vector of the same type as respone. 
Will be used only if a model object was provided in the black_box argument.}
}
\value{
list
}
\description{
Add black box predictions to generated dataset
}
