% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{fit_explanation}
\alias{fit_explanation}
\title{Fit white box model to the simulated data.}
\usage{
fit_explanation(live_object, white_box, selection = FALSE,
  maximum_depth = 0)
}
\arguments{
\item{live_object}{List return by simulate_similar function.}

\item{white_box}{String, learner name recognized by mlr package.}

\item{selection}{If TRUE, variable selection based on AIC will be performed.}

\item{maximum_depth}{maximum depth of a tree (when decision tree is used).}
}
\value{
mlr object returned by train function.
}
\description{
Fit white box model to the simulated data.
}
\examples{
\dontrun{
fitted_explanation <- fit_explanation(local_exploration1, "regr.lm", selection = TRUE)
}

}
