% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{plot_explanation}
\alias{plot_explanation}
\title{Plotting white box models.}
\usage{
plot_explanation(white_box, explained_instance = NULL)
}
\arguments{
\item{white_box}{object returned by mlr::train function.}

\item{explained_instance}{Observation around which model was fitted.
Needed only if forest plot is drawn.}
}
\value{
plot
}
\description{
Plotting white box models.
}
\examples{
\dontrun{
# Forest plot for regression
plot_explanation(fitted_explanation1, winequality_red[5, ])
# Plot decision tree
plot_explanation(fitted_explanation2)
}

}
