% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{plot_explanation}
\alias{plot_explanation}
\title{Plotting white box models.}
\usage{
plot_explanation(white_box, observation = NULL)
}
\arguments{
\item{white_box}{object returned by mlr::train function.}

\item{observation}{Observation around which model was fitted.
Needed only if forest plot is drawn.}
}
\value{
plot
}
\description{
Plotting white box models.
}
\examples{
\dontrun{
dataset_for_local_exploration <- sample_locally(data = winequality_red,
                                               explained_instance = winequality_red[5, ], 
                                               black_box = "regr.svm", 
                                               explained_var = "quality", 
                                               size = 50,
                                               standardise = TRUE)
fitted_explanation1 <- fit_explanation(dataset_for_local_exploration, "regr.lm", selection = TRUE)
fitted_explanation2 <- fit_explanation(dataset_for_local_exploration, "regr.ctree")
# Forest plot for regression
plot_explanation(fitted_explanation1, winequality_red[5, ])
}
# Plot decision tree
plot_explanation(fitted_explanation2)

}
