% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{plot_explanation}
\alias{plot_explanation}
\title{Plotting white box models.}
\usage{
plot_explanation(white_box, observation = NULL)
}
\arguments{
\item{white_box}{object returned by mlr::train function.}

\item{observation}{Observation around which model was fitted.}
}
\value{
plot
}
\description{
Plotting white box models.
}
\examples{
plotwhite_box()

}
