% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_interface.R
\name{plot_explanation}
\alias{plot_explanation}
\title{Plotting white box models.}
\usage{
plot_explanation(model, regr_plot_type = NULL, explained_instance = NULL)
}
\arguments{
\item{model}{object returned by mlr::train function.}

\item{regr_plot_type}{Chr, "forestplot" or "waterfallplot" depending
on which type of plot is to be created.
if lm/glm model is used as interpretable approximation.}

\item{explained_instance}{Observation around which model was fitted.
Needed only if waterfall plot is drawn.}
}
\value{
plot (ggplot2 or base)
}
\description{
DEPRECATED. Please refer to generic plot function for updated and improved interface.
This function will be removed in the future and was left only
to remain consistent with the examples given in https://arxiv.org/abs/1804.01955.
For more, see NEWS and vignette.
}
\examples{
\dontrun{
# Forest plot for regression
plot_explanation(fitted_explanation1, "forestplot", wine[5, ])
# Waterfall plot
plot_explanation(fitted_explanation1, "waterfallplot", wine[5, ])
# Plot decision tree
plot_explanation(fitted_explanation2)
}

}
