% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_explanation2}
\alias{plot_explanation2}
\title{Plotting white box models.}
\usage{
plot_explanation2(explained_model, regr_plot_type = NULL, scale = "logit")
}
\arguments{
\item{explained_model}{List returned by fit_explanation function.}

\item{regr_plot_type}{Chr, "forest" or "waterfall" depending
on which type of plot is to be created.
if lm/glm model is used as interpretable approximation.}

\item{scale}{When probabilities are predicted, they can be plotted on "logit" scale 
or "probability" scale.}
}
\value{
plot (ggplot2 or base)
}
\description{
Plotting white box models.
}
\examples{
\dontrun{
# Forest plot for regression
plot_explanation(fitted_explanation1, "forest", wine[5, ])
# Waterfall plot
plot_explanation(fitted_explanation1, "waterfall", wine[5, ])
# Plot decision tree
plot_explanation(fitted_explanation2)
}

}
