% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_interface.R
\name{plot_regression}
\alias{plot_regression}
\title{Waterfall plot or forestplot for lm/glm explanations.}
\usage{
plot_regression(plot_type, fitted_model, explained_instance, classif)
}
\arguments{
\item{plot_type}{Chr, "forestplot" or "waterfallplot" depending
on which type of plot is to be created.}

\item{fitted_model}{glm or lm object.}

\item{explained_instance}{Observation around which model was fitted.}

\item{classif}{logical, if TRUE, probabilities will be plotted}
}
\value{
plot (ggplot2 or lattice)
}
\description{
DEPRECATED
}
