% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotWhiteBox}
\alias{plotWhiteBox}
\title{Plotting white box models.}
\usage{
plotWhiteBox(trainedModel)
}
\arguments{
\item{trainedModel}{glm object or ctree}
}
\value{
plot
}
\description{
Plotting white box models.
}
