% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{prepare_forestplot}
\alias{prepare_forestplot}
\title{Draw a forest plot with proper annotations.}
\usage{
prepare_forestplot(model)
}
\arguments{
\item{model}{lm/glm object}
}
\description{
Draw a forest plot with proper annotations.
}
