% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{sample_locally}
\alias{sample_locally}
\title{Generate dataset for local exploration.}
\usage{
sample_locally(data, explained_instance, explained_var, size,
  standardise = FALSE)
}
\arguments{
\item{data}{Data frame from which new dataset will be simulated.}

\item{explained_instance}{One row data frame with the same variables
as in data argument. Local exploration will be performed around this observation.}

\item{explained_var}{Name of a column with the variable to be predicted.}

\item{size}{Number of observations is a simulated dataset.}

\item{standardise}{If TRUE, numerical variables will be scaled to have mean 0, var 1.}
}
\value{
list
}
\description{
Generate dataset for local exploration.
}
\examples{
\dontrun{
dataset_for_local_exploration <- sample_locally(data = wine,
                                               explained_instance = wine[5, ],
                                               explained_var = "quality",
                                               size = 50,
                                               standardise = TRUE)
}

}
