% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatesimilar.R
\name{trainWhiteBox}
\alias{trainWhiteBox}
\title{Fit white box model to the simulated data.}
\usage{
trainWhiteBox(liveObject, whiteBox, ...)
}
\arguments{
\item{liveObject}{List return by simulateSimilar function.}

\item{whiteBox}{String, "reg" for linear regression or "dtree" for decision tree.}

\item{...}{Additional arguments passedto makeLearner function.}
}
\value{
mlr object returned by train function.
}
\description{
Fit white box model to the simulated data.
}
