% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatesimilar.R
\name{trainWhiteBox}
\alias{trainWhiteBox}
\title{Fit white box model to the simulated data.}
\usage{
trainWhiteBox(liveObject)
}
\arguments{
\item{liveObject}{}
}
\value{

}
\description{
Fit white box model to the simulated data.
}
